Iss 22 (#29)
* Added Start and End Screen to RoadRun

The user is presented with a start and end screen for a short time
before and after the game starts, displaying the difficulty level and
end of game score.

* Added Titles to Menu Screen

Now we can see what menu we are on. Main menu or choosing a difficulty
for options.

* Added Instruction to Quit Game

The user now sees a message that instructs them to press q if they wish
to quit. This is on the info_win window part of the game screen.

* Update play intructions